test(e2e): fix dashboard heading collision and dedupe onboarding helpers (PER-156)
CI follow-up for PR #116:

- auth-routes smoke broke because the new /dashboard body now renders an
  <h1>Dashboard</h1> alongside the SiteHeader title, so
  getByRole("heading", { name: "Dashboard" }) matched two elements. Scope the
  assertion to the header chrome (its original intent).
- Extract the shared signup→onboard flow into support/onboarding.ts and consume
  it from both dashboard.e2e.ts and budgets.e2e.ts, removing the duplicated
  block that tripped SonarCloud's new-code duplication gate (4.3% > 3%).
